Adds new types structure
