+++ packages/auth/src/types/aws-plugins/cognito-plugin/types/models/authFlowType.ts
@@ -1,0 +1,5 @@
+export type AuthFlowType = 
+	| 'USER_SRP_AUTH'
+	| 'CUSTOM_WITH_SRP'
+	| 'CUSTOM_WITHOUT_SRP'
+	| 'USER_PASSWORD_AUTH'
