+++ packages/auth/src/types/aws-plugins/cognito-plugin/types/models/clientMetadata.ts
@@ -1,0 +1,3 @@
+export type ClientMetadata = {
+	[key: string]: string;
+};
