+++ packages/auth/src/types/aws-plugins/cognito-plugin/types/models/iCognitoStorage.ts
@@ -1,0 +1,6 @@
+export type ICognitoStorage = {
+	setItem(key: string, value: string): void;
+	getItem(key: string): string | null;
+	removeItem(key: string): void;
+	clear(): void;
+}
