+++ packages/auth/src/types/aws-plugins/cognito-plugin/types/models/iCookieStorageData.ts
@@ -1,0 +1,7 @@
+export type ICookieStorageData = {
+	domain: string;
+	path?: string;
+	expires?: number;
+	secure?: boolean;
+	sameSite?: 'string' | 'lax' | 'none';
+}
