+++ packages/auth/src/types/aws-plugins/cognito-plugin/types/models/index.ts
@@ -1,0 +1,14 @@
+import { AuthFlowType } from './authFlowType';
+import { ClientMetadata } from './clientMetadata';
+import { ICognitoStorage } from './ICognitoStorage';
+import { ICookieStorageData } from './iCookieStorageData';
+import { ValidationData } from './validationData';
+
+
+export {
+	ClientMetadata,
+	ValidationData,
+	ICookieStorageData,
+	ICognitoStorage,
+	AuthFlowType
+}
