+++ packages/auth/src/types/aws-plugins/cognito-plugin/types/models/validationData.ts
@@ -1,0 +1,3 @@
+export type ValidationData = {
+	[key: string]: string;
+};
