+++ packages/auth/src/types/aws-plugins/cognito-plugin/types/options/autoSignInOptions.ts
@@ -1,0 +1,5 @@
+import { CognitoSignInOptions } from ".";
+
+export type AutoSignInOptions = {
+	enabled: CognitoSignInOptions;
+}
