+++ packages/auth/src/types/aws-plugins/cognito-plugin/types/options/cognitoSignInOptions.ts
@@ -1,0 +1,6 @@
+import { AuthFlowType, ClientMetadata } from '../models';
+
+export type CognitoSignInOptions = {
+	authFlowType: AuthFlowType;
+	clientMetadata?: ClientMetadata
+}
