+++ packages/auth/src/types/aws-plugins/cognito-plugin/types/options/cognitoSignUpOptions.ts
@@ -1,0 +1,6 @@
+import { ClientMetadata, ValidationData } from '../models'
+
+export type CognitoSignUpOptions = {
+	validationData?: ValidationData;
+	clientMetaData?: ClientMetadata;
+}
