+++ packages/auth/src/types/aws-plugins/cognito-plugin/types/options/index.ts
@@ -1,0 +1,9 @@
+import { AutoSignInOptions } from './autoSignInOptions';
+import { CognitoSignInOptions } from './cognitoSignInOptions';
+import { CognitoSignUpOptions } from './cognitoSignUpOptions';
+
+export {
+	AutoSignInOptions,
+	CognitoSignInOptions,
+	CognitoSignUpOptions
+}
