+++ packages/auth/src/types/models/additionalInfo.ts
@@ -1,0 +1,3 @@
+export type AdditionalInfo = {
+	[key: string]: string;
+}
