+++ packages/auth/src/types/models/authCodeDeliveryDetails.ts
@@ -1,0 +1,14 @@
+import { AuthUserAttributeKey } from '.';
+
+
+export type AuthCodeDeliveryDetails<UserAttributeKey extends AuthUserAttributeKey> = {
+	destination: string;
+	deliveryMedium: DeliveryMedium;
+	attributeName?: UserAttributeKey;
+}
+
+export enum DeliveryMedium {
+	EMAIL = 'EMAIL',
+	SMS = 'SMS',
+	UNKNOWN = 'UNKNOWN'
+}
