+++ packages/auth/src/types/models/authPluginOptions.ts
@@ -1,0 +1,1 @@
+export type AuthPluginOptions = any;
