+++ packages/auth/src/types/models/authSignUpStep.ts
@@ -1,0 +1,4 @@
+export const enum AuthSignUpStep {
+	CONFIRM_SIGN_UP = 'CONFIRM_SIGN_UP',
+	DONE = 'DONE'
+}
