+++ packages/auth/src/types/models/authUserAttribute.ts
@@ -1,0 +1,6 @@
+import { AuthUserAttributeKey } from '.';
+
+export type AuthUserAttribute<UserAttributeKey extends AuthUserAttributeKey = AuthUserAttributeKey> = {
+	userAttributeKey: UserAttributeKey;
+	value: string;
+};
