+++ packages/auth/src/types/models/authUserAttributeKey.ts
@@ -1,0 +1,20 @@
+export type AuthUserAttributeKey = AuthStandardAttributeKey | string;
+
+type AuthStandardAttributeKey = 
+| 'address'
+| 'birthday'
+| 'email'
+| 'family_name'
+| 'gender'
+| 'given_name'
+| 'locale'
+| 'middle_name'
+| 'name'
+| 'nickname'
+| 'phone_number'
+| 'picture'
+| 'preferred_username'
+| 'profile'
+| 'update_at'
+| 'website'
+| 'zoneinfo';
