+++ packages/auth/src/types/models/index.ts
@@ -1,0 +1,24 @@
+import { AdditionalInfo } from './additionalInfo';
+import { AuthCodeDeliveryDetails, DeliveryMedium } from './authCodeDeliveryDetails';
+import { AuthNextSignInStep } from './authNextSignInStep';
+import { AuthNextSignUpStep } from './authNextSignUpStep';
+import { AuthPluginOptions } from './authPluginOptions';
+import { AuthProvider } from './AuthProvider';
+import { AuthSignInStep } from './authSignInStep';
+import { AuthSignUpStep } from './authSignUpStep';
+import { AuthUserAttribute } from './authUserAttribute';
+import { AuthUserAttributeKey } from './authUserAttributeKey';
+
+export {
+	AuthUserAttributeKey,
+	AuthCodeDeliveryDetails,
+	DeliveryMedium,
+	AuthProvider,
+	AuthUserAttribute,
+	AuthPluginOptions,
+	AuthNextSignUpStep,
+	AuthSignUpStep,
+	AdditionalInfo,
+	AuthNextSignInStep,
+	AuthSignInStep
+}
