+++ packages/auth/src/types/options/index.ts
@@ -1,0 +1,6 @@
+import { AuthSignUpOptions } from './AuthSignUpOptions';
+
+
+export {
+	AuthSignUpOptions
+}
