+++ packages/auth/src/types/request/index.ts
@@ -1,0 +1,6 @@
+import { SignUpRequest } from './signUpRequest';
+
+
+export {
+	SignUpRequest
+}
