+++ packages/auth/src/types/result/authSignInResult.ts
@@ -1,0 +1,6 @@
+import { AuthNextSignInStep, AuthUserAttributeKey } from '../models';
+
+export type AuthSignInResult<UserAttributeKey extends AuthUserAttributeKey = AuthUserAttributeKey> = {
+	isSignedIn: boolean;
+	nextStep: AuthNextSignInStep<UserAttributeKey>;
+}
