+++ packages/auth/src/types/result/authSignUpResult.ts
@@ -1,0 +1,6 @@
+import { AuthNextSignUpStep, AuthUserAttributeKey } from '../models';
+
+export type AuthSignUpResult<UserAttributeKey extends AuthUserAttributeKey> = {
+	isSignUpComplete: boolean;
+	nextStep: AuthNextSignUpStep<UserAttributeKey>
+}
