+++ packages/auth/src/types/result/index.ts
@@ -1,0 +1,7 @@
+import { AuthSignInResult } from './authSignInResult';
+import { AuthSignUpResult } from './authSignUpResult';
+
+export {
+	AuthSignUpResult,
+	AuthSignInResult
+}
